New amélioration

+++ src/Controller/UserController.php
@@ -5,29 +5,30 @@ namespace App\Controller;
 use App\Entity\User;
 use App\Form\UserType;
 use App\Repository\UserRepository;
+use PhpParser\Node\Scalar\String_;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
 
-/**
- * @Route("/user")
- */
+
 class UserController extends AbstractController
 {
     /**
-     * @Route("/", name="user_index", methods={"GET"})
+     * @Route("/user/", name="user_index", methods={"GET"})
      */
     public function index(UserRepository $userRepository): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
     }
 
     /**
-     * @Route("/new", name="user_new", methods={"GET","POST"})
+     * @Route("/register", name="user_new", methods={"GET","POST"})
      */
     public function new(Request $request, UserPasswordEncoderInterface $passwordEncoder): Response
     {
@@ -40,7 +41,11 @@ class UserController extends AbstractController
             $user->setPassword($passwordEncoder->encodePassword($user, $user->getPassword()));
             $entityManager->persist($user);
             $entityManager->flush();
-            return $this->redirectToRoute('user_index');
+
+            $id = $user->getId();
+//            return $this->redirectToRoute('user_index');
+            return $this->redirectToRoute('user_show', array('id' => $id));
+
         }
 
         return $this->render('user/new.html.twig', [
@@ -50,27 +55,47 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="user_show", methods={"GET"})
+     * @Route("/user/{id}", name="user_show", methods={"GET"})
      */
-    public function show(User $user): Response
+    public function show(AuthorizationCheckerInterface $authChecker,User $user): Response
     {
+
+
+        if ($user==$this->getUser()){
+
+        $this->denyAccessUnlessGranted('ROLE_USER');
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
+        }
+        elseif ($authChecker->isGranted('ROLE_ADMIN')){
+            return $this->render('user/show.html.twig', [
+                'user' => $user,
+            ]);
+        }
+            return $this->redirectToRoute('home');
+
+
     }
 
     /**
-     * @Route("/{id}/edit", name="user_edit", methods={"GET","POST"})
+     * @Route("/user/{id}/edit", name="user_edit", methods={"GET","POST"})
      */
-    public function edit(Request $request, User $user): Response
+    public function edit(Request $request, User $user,  UserPasswordEncoderInterface $passwordEncoder): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('user_index');
+            $entityManager = $this->getDoctrine()->getManager();
+            $user->setPassword($passwordEncoder->encodePassword($user, $user->getPassword()));
+            $entityManager->persist($user);
+            $entityManager->flush();
+            $id = $user->getId();
+//            return $this->redirectToRoute('user_index');
+            return $this->redirectToRoute('user_show', array('id' => $id));
         }
 
         return $this->render('user/edit.html.twig', [
@@ -80,10 +105,11 @@ class UserController extends AbstractController
     }
 
     /**
-     * @Route("/{id}", name="user_delete", methods={"DELETE"})
+     * @Route("/user/{id}", name="user_delete", methods={"DELETE"})
      */
     public function delete(Request $request, User $user): Response
     {
+        $this->denyAccessUnlessGranted('ROLE_USER');
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $entityManager = $this->getDoctrine()->getManager();
             $entityManager->remove($user);
